Some code organization and cleanup in StripeCheckoutForm
Moving helper functions out for cleanliness, adding initial styles to
form elements
