+++ src/helpers/stripeHelpers.js
@@ -1,0 +1,37 @@
+/* Stripe Helper Functions */
+
+/* Check if we have a zero decimal currency
+ * https://stripe.com/docs/currencies#zero-decimal */
+function zeroDecimalCurrency(amount, currency) {
+  let numberFormat = new Intl.NumberFormat(['en-US'], {
+    style: 'currency',
+    currency: currency,
+    currencyDisplay: 'symbol',
+  });
+  const parts = numberFormat.formatToParts(amount);
+  let zeroDecimalCurrency = true;
+  for (let part of parts) {
+    if (part.type === 'decimal') {
+      zeroDecimalCurrency = false;
+    }
+  }
+  return zeroDecimalCurrency;
+}
+
+/* Format amount for diplay in the UI */
+export function formatAmount(amount, currency) {
+  amount = zeroDecimalCurrency(amount, currency)
+    ? amount
+    : (amount / 100).toFixed(2);
+  return new Intl.NumberFormat('en-US', {
+    style: 'currency',
+    currency,
+  }).format(amount);
+}
+
+/* Format amount for Stripe */
+export function formatAmountForStripe(amount, currency) {
+  return zeroDecimalCurrency(amount, currency)
+    ? amount
+    : Math.round(amount * 100);
+}
